+++ libft/ft_lstsize.c
@@ -6,7 +6,7 @@
 /*   By: nedebies <nedebies@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/12 15:51:35 by nedebies          #+#    #+#             */
-/*   Updated: 2021/04/12 15:51:36 by nedebies         ###   ########.fr       */
+/*   Updated: 2021/04/28 12:30:49 by nedebies         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (lst == NULL)
